backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
pydantic>=2.6.4
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
email-validator>=2.2.0
cryptography>=42.0.8
requests>=2.31.0
